Change to a message-based approach
The previous version of `async with` used a dependency graph to
establish the order of execution of the asynchronous tasks. This
approach covered most cases, but it had one flaw: the vertices
of the dependency graph could not have common out-neighbors. This
constraint was a design decision, imposed so the dependency graph
could be translated to a series of nested tasks (Elixir's Task
module). That's why, at some point in the process, vertices with
common out-neighbors were merged into a single vertex. Here is an
example:

           A                        A
          ↙ ↘                      ↙ ↘
         C   D     B      =>      C   D,B
        ↙ ↘   ↘  ↙               ↙ ↘    ↘
       E   F    G               E   F     G

However, this approach did not cover some edge cases.

    async with a <- 1,
               b <- :timer.sleep(1000),
               c <- IO.inspect(a),
               d <- {a, b} do
    end

In the example above, "1" should be printed immediately because
C only has a dependency on A and is unrelated to B. But, it was
not. This was due to the resulting dependency graph, in which
both A and B were executed in the same step.

           A    B                 A,B
          ↙ ↘ ↙         =>       ↙  ↘
         C   D                  C    D

To handle these scenarios, a new approach was proposed in Elixir
Core Mailing list: to ditch the dependency graph and use message
passing to solve the dependencies between clauses. These are the
steps of the new approach:

  1. Spawn a process per each clause whose dependencies are
     processed.
  2. Wait for replies.
  3. Process the reply:
      3.1. Remove the clause from the list of clauses to be
           processed.
      3.2. Update the current list of values
  4. Repeat step 1 until all the clauses are processed

Here is an example:

    async with a <- 1,
               b <- a + 2,
               c <- {a, b} do
    end

As you can see, there are three clauses:

  * `a <- 1` with no dependencies
  * `b <- a + 2` with a dependency on the variable `a`
  * `c <- {a, b}` with a dependency in both variables `a` and `b`

The final sequence would be then:

   1. Initial state:
        - Clauses: [`a <- 1`, `b <- a + 2`, `c <- {a, b}`]
        - Values: []
   2. Spawn `a <- 1` as it does not have dependencies.
   3. Wait for `a <- 1` reply.
   4. Process `a <- 1` reply:
        - Clauses: [`b <- a + 2`, `c <- {a, b}`]
        - Values: [a: 1]
   5. Spawn `b <- a + 2` now that `a` is processed.
   6. Wait for `b <- a + 2` reply.
   7. Process `b <- a + 2` reply:
        - Clauses: [`c <- {a, b}`]
        - Values: [a: 1, b: 3]
   8. Spawn `c <- {a, b}` now that `a` and `b` are processed.
   9. Wait for `c <- {a, b}` reply.
  10. Process `c <- {a, b}` reply:
        - Clauses: []
        - Values: [a: 1, b: 3, c: {1, 3}]
  11. Return [a: 1, b: 3, c: {1, 3}]

In case of failure, the error is returned immediately and all
the processes are killed.