Refactor Clause into Clauses
Clauses are now defined by a function and its defined and used
vars. The Clause struct is no longer present, as it was not useful
when generating the new macros. The resulting map is a 1-1
conversion between the macro-generation code and the macro code.

Var renamings are version-based instead of timestamp-based. This
should provide better performance and predictability.