Replace "not" with "!" to be backwards compatible
The new formatter introduced in Elixir v1.6 changes "not x in y"
expressions to "x not in y". The "not in" syntaxis was introduced
in Elixir v1.5, and it produces a backwards incompatibility.

Since AsyncWith currently supports v1.4, this is a workaround to
keep the formatter happy.

More info: https://github.com/elixir-lang/elixir/issues/7227