Re-raise unrescued errors
Unrescued errors were producing the termination of the process of
the runner, returning:

    {:exit, {error, <stacktrace>}}

This was a bit confusing, as the expected behavior would be to get
an error raised. Here is an example:

    try do
      async with _ <- raise("ops"), do: :ok
    rescue
      error -> error
    end

The common developer would expect the example above to return
`%RuntimeError{message: "ops"}` and not
`{:exit, {%RuntimeError{message: "ops"}, ...}}`.